+++ CHANGELOG.md
@@ -19,3 +19,8 @@
 - Se realizaron arreglos al NavBar inicial
 - Favicon agregado
 - Se instaló React-Spinners
+
+## ItemListDetail Created
+- Se creó el componente *ItemDetailContainer* que se encarga de hacer un fetch() del item
+- Se creó el componente *ItemDetail* que se encarga de renderizar el item
+- Se creó el componente *ItemCount* dentro de ItemDetail